+++ cumulusci/tasks/salesforce/Deploy.py
@@ -56,11 +56,9 @@ class Deploy(BaseSalesforceMetadataApiTask):
         "transforms": {
             "description": "Apply source transforms before deploying. See the CumulusCI documentation for details on how to specify transforms."
         },
-        "rest_deploy": {
-            "description": "If True, deploy metadata using REST API"
-        },
-        "clean_profiles": {
-            "description": "If specified, all profiles are cleaned of invalid references before deployment."
+        "rest_deploy": {"description": "If True, deploy metadata using REST API"},
+        "clean_invalid_ref": {
+            "description": "If specified, all profiles and permission sets are cleaned of invalid references before deployment."
         },
     }
