+++ cumulusci/utils/clean_invalid_references.py
